# docs/swagger.yaml
basePath: /
definitions:
...
  iam.AssignRoleRequest:
    properties:
      roleId:
        type: string
    required:
    - roleId
    type: object
  iam.CreateUserRequest:
    properties:
      displayName:
        type: string
      email:
        type: string
      employeeId:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Deep QMS backend application serving the core HR, IT, and specialized
    logic domains.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DML API
  version: "1.0"
paths:
...
  /api/v1/business-units:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of business units for the authenticated
        tenant.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Page size
        in: query
        name: size
        type: integer
      - description: Search term (name/code)
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paginated business unit data
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List business units
      tags:
      - Organization
  /api/v1/business-units/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a specific business unit by its ID.
      parameters:
      - description: Business Unit ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Business unit data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid ID format
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a business unit
      tags:
      - Organization
...
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of users for the authenticated tenant.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Page size
        in: query
        name: size
        type: integer
      - description: Search term (email/name)
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paginated user data
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Creates a new user profile attached to an employee reference.
      parameters:
      - description: User details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/iam.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created user
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request payload
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden (Requires ADMIN)
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new user
      tags:
      - Users
  /api/v1/users/{userID}/roles:
    post:
      consumes:
      - application/json
      description: Binds an existing RBAC role to a specific user.
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Role Assignment specifics
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/iam.AssignRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Role assigned successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request payload
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden (Requires ADMIN)
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Assign role to user
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"